First efforts towards new tree node organization
* Rename long methods into shorter versions
    - Transform Java-ish getters and setters into properties
* Create new core components (LearningNodeMixin, ActiveLeafMixin, InactiveLeafMixin)
    - They aim at simplifying the way we assemble learning nodes
    - TODO: revisit the mixin suffix in the future and evaluate possibly better alternatives for the naming pattern
* Apply new structure to the nodes of HTC
