+++ src/skmultiflow/trees/isoup_tree.py
@@ -489,16 +489,16 @@ class iSOUPTreeRegressor(HoeffdingTreeRegressor, MultiOutputMixin):
             if self._growth_allowed and \
                     isinstance(learning_node, ActiveLearningNode):
                 active_learning_node = learning_node
-                weight_seen = active_learning_node.get_weight_seen()
+                weight_seen = active_learning_node.total_weight
 
                 weight_diff = weight_seen - active_learning_node.\
-                    get_weight_seen_at_last_split_evaluation()
+                    last_split_attempt_at
                 if weight_diff >= self.grace_period:
                     self._attempt_to_split(active_learning_node,
                                            found_node.parent,
                                            found_node.parent_branch)
                     active_learning_node.\
-                        set_weight_seen_at_last_split_evaluation(weight_seen)
+                        last_split_attempt_at = weight_seen
         # Split node encountered a previously unseen categorical value
         # (in a multiway test)
         elif isinstance(leaf_node, SplitNode) and \
@@ -651,8 +651,8 @@ class iSOUPTreeRegressor(HoeffdingTreeRegressor, MultiOutputMixin):
         else:
             hoeffding_bound = self.compute_hoeffding_bound(
                 split_criterion.get_range_of_merit(
-                    node.get_stats()
-                ), self.split_confidence, node.get_weight_seen())
+                    node.stats
+                ), self.split_confidence, node.total_weight)
             best_suggestion = best_split_suggestions[-1]
             second_best_suggestion = best_split_suggestions[-2]
 
@@ -690,7 +690,7 @@ class iSOUPTreeRegressor(HoeffdingTreeRegressor, MultiOutputMixin):
             else:
                 new_split = self.new_split_node(
                     split_decision.split_test,
-                    node.get_stats()
+                    node.stats
                 )
                 for i in range(split_decision.num_splits()):
                     new_child = self._new_learning_node(
