+++ src/skmultiflow/trees/nodes/__init__.py
@@ -3,20 +3,19 @@ The :mod:`skmultiflow.trees.nodes` module includes learning and split node
 implementations for the hoeffding trees.
 """
 
-from .found_node import FoundNode
-from .node import Node
-from .ada_node import AdaNode
-from .split_node import SplitNode
-from .learning_node import LearningNode
-from .htc_nodes import ActiveLearningNode, InactiveLearningNode, ActiveLearningNodeNB, \
-    ActiveLearningNodeNBAdaptive
-from .hatc_nodes import AdaSplitNode, AdaLearningNode
-from .arf_htc_nodes import RandomActiveLearningNode, RandomActiveLearningNodeNB, \
-    RandomActiveLearningNodeNBAdaptive
-from .efdtc_nodes import EFDTSplitNode, EFDTActiveLearningNode, EFDTInactiveLearningNode, \
-    EFDTActiveLearningNodeNB, EFDTActiveLearningNodeNBAdaptive
-from .lc_htc_nodes import LCActiveLearningNode, LCInactiveLearningNode, LCLearningNodeNB, \
-    LCLearningNodeNBA
+from .core import FoundNode, Node, SplitNode, LearningNodeMixin, ActiveLeafMixin, \
+    InactiveLeafMixin, AdaNode
+from .htc_nodes import ActiveLearningNodeMC, InactiveLearningNodeMC, ActiveLearningNodeNB, \
+    ActiveLearningNodeNBA
+from .hatc_nodes import AdaSplitNode, AdaLearningNode  # TODO: verify name
+from .arf_htc_nodes import RandomActiveLearningNodeMC, RandomActiveLearningNodeNB, \
+    RandomActiveLearningNodeNBA
+from .efdtc_nodes import EFDTSplitNode, EFDTActiveLearningNodeMC, EFDTInactiveLearningNodeMC, \
+    EFDTActiveLearningNodeNB, EFDTActiveLearningNodeNBA
+from .lc_htc_nodes import LCActiveLearningNodeMC, LCInactiveLearningNodeMC, \
+    LCActiveLearningNodeNB, LCActiveLearningNodeNBA
+
+# TODO continue from here
 from .htr_nodes import ActiveLearningNodeForRegression, ActiveLearningNodePerceptron, \
     InactiveLearningNodeForRegression, InactiveLearningNodePerceptron
 from .arf_htr_nodes import RandomLearningNodeForRegression, RandomLearningNodePerceptron
@@ -28,19 +27,20 @@ from .sst_htr_nodes import SSTActiveLearningNode, SSTActiveLearningNodeAdaptive,
     SSTInactiveLearningNode, SSTInactiveLearningNodeAdaptive
 
 
-__all__ = ["FoundNode", "Node", "SplitNode", "LearningNode", "ActiveLearningNode",
-           "InactiveLearningNode", "ActiveLearningNodeNB", "ActiveLearningNodeNBAdaptive",
-           "RandomActiveLearningNode", "RandomActiveLearningNodeNB",
-           "RandomActiveLearningNodeNBAdaptive", "AdaNode", "AdaSplitNode", "AdaLearningNode",
-           "EFDTSplitNode", "EFDTActiveLearningNode", "EFDTInactiveLearningNode",
-           "EFDTActiveLearningNodeNB", "EFDTActiveLearningNodeNBAdaptive",
-           "LCActiveLearningNode", "LCInactiveLearningNode", "LCLearningNodeNB",
-           "LCLearningNodeNBA", "ActiveLearningNodeForRegression", "ActiveLearningNodePerceptron",
-           "InactiveLearningNodeForRegression", "InactiveLearningNodePerceptron",
-           "RandomLearningNodeForRegression", "RandomLearningNodePerceptron",
-           "AdaSplitNodeForRegression", "AdaLearningNodeForRegression",
-           "ActiveLearningNodeForRegressionMultiTarget", "ActiveLearningNodePerceptronMultiTarget",
-           "ActiveLearningNodeAdaptiveMultiTarget", "InactiveLearningNodePerceptronMultiTarget",
-           "InactiveLearningNodeAdaptiveMultiTarget", "SSTActiveLearningNode",
-           "SSTActiveLearningNodeAdaptive", "SSTInactiveLearningNode",
+__all__ = ["FoundNode", "Node", "LearningNodeMixin", "ActiveLeafMixin", "InactiveLeafMixin",
+           "SplitNode", "ActiveLearningNodeMC", "InactiveLearningNodeMC",
+           "ActiveLearningNodeNB", "ActiveLearningNodeNBA",
+           "RandomActiveLearningNodeMC", "RandomActiveLearningNodeNB",
+           "RandomActiveLearningNodeNBA", "AdaNode", "AdaSplitNode", "AdaLearningNode",
+           "EFDTSplitNode", "EFDTActiveLearningNodeMC", "EFDTInactiveLearningNodeMC",
+           "EFDTActiveLearningNodeNB", "EFDTActiveLearningNodeNBA",
+           "LCActiveLearningNodeMC", "LCInactiveLearningNodeMC", "LCActiveLearningNodeNB",
+           "LCActiveLearningNodeNBA", "ActiveLearningNodeForRegression",
+           "ActiveLearningNodePerceptron", "InactiveLearningNodeForRegression",
+           "InactiveLearningNodePerceptron", "RandomLearningNodeForRegression",
+           "RandomLearningNodePerceptron", "AdaSplitNodeForRegression",
+           "AdaLearningNodeForRegression", "ActiveLearningNodeForRegressionMultiTarget",
+           "ActiveLearningNodePerceptronMultiTarget", "ActiveLearningNodeAdaptiveMultiTarget",
+           "InactiveLearningNodePerceptronMultiTarget", "InactiveLearningNodeAdaptiveMultiTarget",
+           "SSTActiveLearningNode", "SSTActiveLearningNodeAdaptive", "SSTInactiveLearningNode",
            "SSTInactiveLearningNodeAdaptive"]
